requirements.txt
PyYAML~=5.4.1
typer~=0.4.0
Pillow~=8.2.0
PyPDF2~=1.26.0